% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cir.R
\name{areacir}
\alias{areacir}
\title{Area of a circle}
\usage{
areacir(x)
}
\arguments{
\item{x}{The radius of the intended circle}
}
\value{
The output from \code{\link{print}}
}
\description{
Area of a circle
}
\examples{
areacir ("2")
}
