% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rec.R
\name{arearec}
\alias{arearec}
\title{Area of rectangle}
\usage{
arearec(x, y)
}
\arguments{
\item{x}{The length of the rectangle}

\item{y}{The breath of the rectangle}
}
\value{
The output from \code{\link{print}}
}
\description{
Area of rectangle
}
\examples{
arearec ("5,8")
}
