% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tri.R
\name{areatri}
\alias{areatri}
\title{Area of a triangle}
\usage{
areatri(x, y)
}
\arguments{
\item{x}{The base of the triangle}

\item{y}{The height of the triangle}
}
\value{
The output from \code{\link{print}}
}
\description{
Area of a triangle
}
\examples{
areatri(2,4)
}
