% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{filter_on_A}
\alias{filter_on_A}
\title{Example of using dplyr}
\usage{
filter_on_A(data, p)
}
\arguments{
\item{data}{A data frame containing a column named A}

\item{p}{A predicate function}
}
\value{
The data frame filtered to those rows where p is true on A
}
\description{
Example of using dplyr
}
